refactor: move some utils to common utils
