+++ extension/common/utils/gobject.ts
